Revert "Possible Safari Fix"
This reverts commit e828da94e5794b14cb8369dc2a03b854bafc0fcd.

Safari blocks cookies that are Third Party by default. This comes from
the request being corss origin. There is no good solution to this other
then having users disable this setting or setting up a domain to use for
both the API and frontend which is the solution I would probably opt for
if this was an actual application.